+++ lib/x402/payments.rb
@@ -16,37 +16,77 @@ require_relative "payments/v2/headers"
 require_relative "payments/v2/payload_builder"
 require_relative "payments/response_parser"
 require_relative "payments/generator"
+require_relative "payments/solana/generator"
 
 module X402
   module Payments
 
     class << self
-      def generate_header(amount:, resource:, description: nil, network: nil, private_key: nil, pay_to: nil, extra: nil, version: nil)
-        generator = Generator.new
-        generator.generate_header(
-          amount: amount,
-          resource: resource,
-          description: description,
-          network: network,
-          private_key: private_key,
-          pay_to: pay_to,
-          extra: extra,
-          version: version
-        )
+      def generate_header(amount:, resource:, description: nil, network: nil, private_key: nil, pay_to: nil, extra: nil, version: nil, fee_payer: nil)
+        chain_name = network || configuration.chain
+
+        if solana_chain?(chain_name)
+          generator = Solana::Generator.new
+          generator.generate_header(
+            amount: amount,
+            resource: resource,
+            description: description,
+            network: network,
+            private_key: private_key,
+            pay_to: pay_to,
+            version: version,
+            fee_payer: fee_payer
+          )
+        else
+          generator = Generator.new
+          generator.generate_header(
+            amount: amount,
+            resource: resource,
+            description: description,
+            network: network,
+            private_key: private_key,
+            pay_to: pay_to,
+            extra: extra,
+            version: version
+          )
+        end
       end
 
-      def generate_link(amount:, resource:, description: nil, network: nil, private_key: nil, pay_to: nil, extra: nil, version: nil)
-        generator = Generator.new
-        generator.generate_link(
-          amount: amount,
-          resource: resource,
-          description: description,
-          network: network,
-          private_key: private_key,
-          pay_to: pay_to,
-          extra: extra,
-          version: version
-        )
+      def generate_link(amount:, resource:, description: nil, network: nil, private_key: nil, pay_to: nil, extra: nil, version: nil, fee_payer: nil)
+        chain_name = network || configuration.chain
+        protocol_version = version || configuration.protocol_version
+        header_name = payment_header_name(protocol_version)
+
+        if solana_chain?(chain_name)
+          generator = Solana::Generator.new
+          header = generator.generate_header(
+            amount: amount,
+            resource: resource,
+            description: description,
+            network: network,
+            private_key: private_key,
+            pay_to: pay_to,
+            version: protocol_version,
+            fee_payer: fee_payer
+          )
+          {
+            payment_header: header,
+            header_name: header_name,
+            curl_command: "curl -s -H \"#{header_name}: #{header}\" #{resource} | jq ."
+          }
+        else
+          generator = Generator.new
+          generator.generate_link(
+            amount: amount,
+            resource: resource,
+            description: description,
+            network: network,
+            private_key: private_key,
+            pay_to: pay_to,
+            extra: extra,
+            version: version
+          )
+        end
       end
 
       def generate_header_for(payment_required, private_key: nil)
@@ -91,6 +131,7 @@ module X402
           nil
         end
       end
+
     end
   end
 end